ann/requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
gradio>=4.0.0
pillow>=10.0.0
